# Improve CompositeByteBuf sequential access performance