Motivation:

While running some other benchmarks I noticed that sequential access of `CompositeByteBuf`s is very slow, particularly if they are large.
 
Modifications:

- Override `forEachByte(...)` methods
- Remove redundant `Component.length` field and standardize component offset modifications via `setOffset()`
- Simplify the general pattern used where a sub-range is iterated over
- Add racy cache of last-accessed Component to `findComponent(int)` method
- Add benchmark to demonstrate improvement

The cache is based on the fact that a call to `findComponent` precededs every individual read/write access, and the assumption that very often the same component will be needed between successive accesses.
 
Result:

I will paste benchmark results below. Also some anecdotal observations:
- There is a "cache hit" rate of 75% when running the existing buffer package unit tests
- The complete buffer package unit tests appear to take 30% less time to run if `AbstractByteBufTest.CAPACITY` is increased by a factor of 10

In addition to the raw speed there should also be a minor mem/garbage benefit due to removal of `Component.length`.