@njhill 2018-10-18
Preliminary "quick" benchmark results (java8, linux x86_64), I will run longer ones and update this soon.

Before:
```
Benchmark                                                   (bufferType)   (size)   Mode  Cnt         Score          Error  Units
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS        8  thrpt    4  22616169.666 ±   176581.796  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS       64  thrpt    4  20090370.077 ±  1504256.400  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS     1024  thrpt    4  16439870.879 ±  1869073.291  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS    10240  thrpt    4  13709075.845 ±   335632.537  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS   102400  thrpt    4  11433894.168 ±  1775238.013  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS  1024000  thrpt    4   9230941.943 ±  1321042.676  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS        8  thrpt    4  26295411.342 ± 15845070.362  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS       64  thrpt    4  25861547.001 ± 16899224.648  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS     1024  thrpt    4  23655742.901 ±  8111894.373  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS    10240  thrpt    4  19977620.847 ±  4864887.439  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS   102400  thrpt    4  14504533.962 ± 15675813.915  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS  1024000  thrpt    4  12705298.535 ±  7115391.330  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS        8  thrpt    4   3285759.893 ±  1213558.650  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS       64  thrpt    4    400081.408 ±    87590.973  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS     1024  thrpt    4     12241.967 ±     1538.719  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS    10240  thrpt    4       922.509 ±       87.584  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS   102400  thrpt    4        91.756 ±       36.221  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS  1024000  thrpt    4         8.156 ±        0.422  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS        8  thrpt    4   5490764.400 ±   358787.988  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS       64  thrpt    4    756758.233 ±    62124.704  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS     1024  thrpt    4     34610.843 ±     5116.880  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS    10240  thrpt    4      2843.865 ±      132.271  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS   102400  thrpt    4       177.839 ±       10.650  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS  1024000  thrpt    4        13.814 ±        1.071  ops/s
```

After:
```
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS        8  thrpt    4  28047447.834 ± 3940346.622  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS       64  thrpt    4  23705616.720 ± 8710246.761  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS     1024  thrpt    4  20137684.432 ± 2557539.936  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS    10240  thrpt    4  14796841.680 ± 1706886.641  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS   102400  thrpt    4  13891358.610 ±  447864.029  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             SMALL_CHUNKS  1024000  thrpt    4  11003133.524 ± 5697106.106  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS        8  thrpt    4  34823320.750 ± 6989208.081  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS       64  thrpt    4  34627191.013 ± 2488054.197  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS     1024  thrpt    4  29875818.004 ± 7035011.337  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS    10240  thrpt    4  27988096.912 ± 2377740.064  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS   102400  thrpt    4  20492341.893 ± 4551135.628  ops/s
CompositeByteBufSequentialBenchmark.forEachByte             LARGE_CHUNKS  1024000  thrpt    4  18292347.103 ± 2246528.786  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS        8  thrpt    4   4549258.960 ±  427625.203  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS       64  thrpt    4    709602.020 ±  146589.239  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS     1024  thrpt    4     30766.179 ±    4715.836  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS    10240  thrpt    4      2147.270 ±     116.355  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS   102400  thrpt    4       214.431 ±      48.717  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  SMALL_CHUNKS  1024000  thrpt    4        19.393 ±       2.777  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS        8  thrpt    4   9170553.992 ±  985955.240  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS       64  thrpt    4   1330377.778 ±  110892.917  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS     1024  thrpt    4     64933.227 ±   19936.330  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS    10240  thrpt    4      5763.821 ±     531.953  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS   102400  thrpt    4       478.730 ±      46.962  ops/s
CompositeByteBufSequentialBenchmark.sequentialWriteAndRead  LARGE_CHUNKS  1024000  thrpt    4        60.766 ±      21.751  ops/s
```