@njhill 2018-10-18
Updated the above results with longer benchmark run on latest code. Also clarified my comment about the existing unit test speedup - this applied when I had increased the buffer capacity used from 4k to 40k.