@normanmaurer 2018-10-19
@netty-bot test this please