@njhill 2018-10-20
Added an additional commit to avoid slicing the buffers - the `Component`s themselves end up being kind of lightweight slices; to reach a target index one less pointer needs to be followed and only a single sum is needed rather than two (offset within slice followed by slice adjustment to underlying buff). Should also help a lot with garbage. Sorry for churn with the PR, @normanmaurer let me know if you want me to do this in a follow on instead.

New benchmark numbers:

Before:

```
Benchmark               (bufferType)   (size)   Mode  Cnt         Score         Error  Units
forEachByte             SMALL_CHUNKS        8  thrpt   20  23321032.007 ± 1200986.883  ops/s
forEachByte             SMALL_CHUNKS       64  thrpt   20  21121192.567 ±  625995.308  ops/s
forEachByte             SMALL_CHUNKS     1024  thrpt   20  16427286.003 ± 1302257.690  ops/s
forEachByte             SMALL_CHUNKS    10240  thrpt   20  13816112.113 ±  954850.509  ops/s
forEachByte             SMALL_CHUNKS   102400  thrpt   20  12680408.746 ±  511341.185  ops/s
forEachByte             SMALL_CHUNKS  1024000  thrpt   20  10293156.043 ±  499010.112  ops/s
forEachByte             LARGE_CHUNKS        8  thrpt   20  29709271.583 ±  639642.663  ops/s
forEachByte             LARGE_CHUNKS       64  thrpt   20  29976430.650 ±  487665.514  ops/s
forEachByte             LARGE_CHUNKS     1024  thrpt   20  26093744.231 ± 1259157.941  ops/s
forEachByte             LARGE_CHUNKS    10240  thrpt   20  20928532.022 ± 1548531.089  ops/s
forEachByte             LARGE_CHUNKS   102400  thrpt   20  18022356.570 ± 1089768.399  ops/s
forEachByte             LARGE_CHUNKS  1024000  thrpt   20  15847645.984 ±  303401.555  ops/s
sequentialWriteAndRead  SMALL_CHUNKS        8  thrpt   20   3751755.080 ±  240132.581  ops/s
sequentialWriteAndRead  SMALL_CHUNKS       64  thrpt   20    397353.694 ±   13453.313  ops/s
sequentialWriteAndRead  SMALL_CHUNKS     1024  thrpt   20     13177.540 ±    1206.362  ops/s
sequentialWriteAndRead  SMALL_CHUNKS    10240  thrpt   20      1018.208 ±      32.802  ops/s
sequentialWriteAndRead  SMALL_CHUNKS   102400  thrpt   20        89.824 ±       4.091  ops/s
sequentialWriteAndRead  SMALL_CHUNKS  1024000  thrpt   20         7.885 ±       0.322  ops/s
sequentialWriteAndRead  LARGE_CHUNKS        8  thrpt   20   5385825.536 ±  323582.099  ops/s
sequentialWriteAndRead  LARGE_CHUNKS       64  thrpt   20    723478.679 ±   62246.268  ops/s
sequentialWriteAndRead  LARGE_CHUNKS     1024  thrpt   20     31303.200 ±    3527.802  ops/s
sequentialWriteAndRead  LARGE_CHUNKS    10240  thrpt   20      2398.795 ±     202.460  ops/s
sequentialWriteAndRead  LARGE_CHUNKS   102400  thrpt   20       213.001 ±      13.447  ops/s
sequentialWriteAndRead  LARGE_CHUNKS  1024000  thrpt   20        16.606 ±       1.266  ops/s
```

After:

```
Benchmark               (bufferType)   (size)   Mode  Cnt         Score         Error  Units
forEachByte             SMALL_CHUNKS        8  thrpt   20  45155942.036 ± 1659946.309  ops/s
forEachByte             SMALL_CHUNKS       64  thrpt   20  34072026.042 ± 3034790.476  ops/s
forEachByte             SMALL_CHUNKS     1024  thrpt   20  24160151.968 ±  839203.577  ops/s
forEachByte             SMALL_CHUNKS    10240  thrpt   20  17695393.795 ±  637180.547  ops/s
forEachByte             SMALL_CHUNKS   102400  thrpt   20  14701242.149 ±  760724.176  ops/s
forEachByte             SMALL_CHUNKS  1024000  thrpt   20  11710681.429 ±  898633.272  ops/s
forEachByte             LARGE_CHUNKS        8  thrpt   20  52925728.195 ± 3876011.234  ops/s
forEachByte             LARGE_CHUNKS       64  thrpt   20  54907607.479 ±  631410.724  ops/s
forEachByte             LARGE_CHUNKS     1024  thrpt   20  43256221.100 ± 6374400.426  ops/s
forEachByte             LARGE_CHUNKS    10240  thrpt   20  38285444.793 ± 1412667.765  ops/s
forEachByte             LARGE_CHUNKS   102400  thrpt   20  22194272.225 ±  643517.544  ops/s
forEachByte             LARGE_CHUNKS  1024000  thrpt   20  19506279.642 ± 1314325.338  ops/s
sequentialWriteAndRead  SMALL_CHUNKS        8  thrpt   20   5305074.684 ±  183019.676  ops/s
sequentialWriteAndRead  SMALL_CHUNKS       64  thrpt   20    854624.918 ±   64613.004  ops/s
sequentialWriteAndRead  SMALL_CHUNKS     1024  thrpt   20     32992.834 ±    1854.354  ops/s
sequentialWriteAndRead  SMALL_CHUNKS    10240  thrpt   20      2560.711 ±     103.763  ops/s
sequentialWriteAndRead  SMALL_CHUNKS   102400  thrpt   20       239.707 ±      11.058  ops/s
sequentialWriteAndRead  SMALL_CHUNKS  1024000  thrpt   20        20.662 ±       0.450  ops/s
sequentialWriteAndRead  LARGE_CHUNKS        8  thrpt   20  11737082.384 ±  688989.118  ops/s
sequentialWriteAndRead  LARGE_CHUNKS       64  thrpt   20   1610604.950 ±  107840.693  ops/s
sequentialWriteAndRead  LARGE_CHUNKS     1024  thrpt   20     82586.310 ±    6226.571  ops/s
sequentialWriteAndRead  LARGE_CHUNKS    10240  thrpt   20     10004.470 ±    1724.012  ops/s
sequentialWriteAndRead  LARGE_CHUNKS   102400  thrpt   20      1017.058 ±      25.493  ops/s
sequentialWriteAndRead  LARGE_CHUNKS  1024000  thrpt   20       112.308 ±      11.088  ops/s
```